# GTEST/UCP: Disabled performance checking for UD transports.